py3: dvc/utils/compat.py -> dvc/compat.py
