+++ dvc/remote/local.py
@@ -28,7 +28,7 @@ from dvc.utils import makedirs
 from dvc.utils import relpath
 from dvc.utils import tmp_fname
 from dvc.utils import walk_files
-from dvc.utils.compat import fspath_py35
+from dvc.compat import fspath_py35
 from dvc.utils.fs import move
 from dvc.utils.fs import remove
