+++ dvc/state.py
@@ -11,7 +11,7 @@ from dvc.exceptions import DvcException
 from dvc.utils import current_timestamp
 from dvc.utils import relpath
 from dvc.utils import to_chunks
-from dvc.utils.compat import fspath_py35
+from dvc.compat import fspath_py35
 from dvc.utils.fs import get_inode
 from dvc.utils.fs import get_mtime_and_size
 from dvc.utils.fs import remove
