+++ tests/dir_helpers.py
@@ -50,7 +50,7 @@ import pytest
 from funcy.py3 import lmap, retry
 
 from dvc.utils import makedirs
-from dvc.utils.compat import fspath, fspath_py35
+from dvc.compat import fspath, fspath_py35
 
 
 __all__ = ["tmp_dir", "scm", "dvc", "repo_template", "run_copy", "erepo_dir"]
