+++ tests/func/test_get.py
@@ -10,7 +10,7 @@ from dvc.repo.get import GetDVCFileError, PathMissingError
 from dvc.repo import Repo
 from dvc.system import System
 from dvc.utils import makedirs
-from dvc.utils.compat import fspath
+from dvc.compat import fspath
 from tests.utils import trees_equal
